Reworked FirefoxAccount construction
Made `FirefoxAccount` input a single `FxaConfig` instance for all of
it's configuration.  I want to add more constructor arguments but don't
want to end up with a huge argument count.

Made the example code use the public API, not the internal module.
`FxaConfig` helps ease this transition.  The motivation for this is that
I want to move FxaClent and some other fields to the public
`FirefoxAccount` instance, which means the example code needs to start
using it. Deleted the `fxa-client/migration` example, we're not really
using the migration code anymore.

Adding some handling for dead code now that we're not exposing the
`internal` module as pub.

Created an `FxaServer` enum.  This matches how the android wrapper
handles it and I don't think we need to input arbitrary URLs here.